+++ R/simulation.R
@@ -80,6 +80,7 @@ adapt_charging_features <- function (sessions, time_resolution = 15, power_resol
       ConnectionStartDateTime = round_date(.data$ConnectionStartDateTime, paste(time_resolution, "mins")),
       ConnectionEndDateTime = .data$ConnectionStartDateTime + convert_time_num_to_period(.data$ConnectionHours),
       Power = round_to_interval(.data$Power, interval = power_resolution),
+      ConnectionHours = round(as.numeric(.data$ConnectionEndDateTime - .data$ConnectionStartDateTime, unit="hours"), 2),
       ChargingHours = round(pmin(.data$Energy/.data$Power, .data$ConnectionHours), 2),
       Energy = round(.data$Power * .data$ChargingHours, 2),
       ChargingStartDateTime = .data$ConnectionStartDateTime,